+++ src/Tiles/Caching/PrefetchManager.cpp
@@ -68,10 +68,10 @@ void PrefetchManager::BuildDownloadList(BaseProvider* provider, int zoom, CRect 
     provider->get_Projection()->GetTileMatrixMinXY(zoom, size1);
     provider->get_Projection()->GetTileMatrixMaxXY(zoom, size2);
 
-    const int minX = (int)BaseProjection::Clip(indices.left, size1.cx, size2.cx);
-    const int maxX = (int)BaseProjection::Clip(indices.right, size1.cy, size2.cy);
-    const int minY = (int)BaseProjection::Clip(MIN(indices.top, indices.bottom), size1.cx, size2.cx);
-    const int maxY = (int)BaseProjection::Clip(MAX(indices.top, indices.bottom), size1.cy, size2.cy);
+    const int minX = static_cast<int>(BaseProjection::Clip(indices.left, size1.cx, size2.cx));
+    const int maxX = static_cast<int>(BaseProjection::Clip(indices.right, size1.cy, size2.cy));
+    const int minY = static_cast<int>(BaseProjection::Clip(MIN(indices.top, indices.bottom), size1.cx, size2.cx));
+    const int maxY = static_cast<int>(BaseProjection::Clip(MAX(indices.top, indices.bottom), size1.cy, size2.cy));
 
     const int centX = (maxX + minX) / 2;
     const int centY = (maxY + minY) / 2;
@@ -138,7 +138,7 @@ long PrefetchManager::Prefetch(BaseProvider* provider, CRect indices, int zoom, 
     // actual call to do the job
     _loader.Load(points, provider, zoom, info);
 
-    const long size = points.size();
+    const long size = static_cast<long>(points.size());
 
     TilePoint::ReleaseMemory(points);
